Move 'reverseString' to utils.c
